Unsplit lttng events
LTTng has a limitation in that it cannot emit events with more than 9
parameters.  Previously we would split these events into multiple
events.  However this proves to be very difficult to consume with
TraceEvent.  Instead we now pack the parameters into a byte array an
emit the byte array + length as a single event, avoiding the need to
recombine events later.